requirements.txt
requests>=2.31.0
Django==5.0.3
djangorestframework==3.14.0
python-dotenv==1.0.1
django-cors-headers==4.3.1
pytest==8.0.0
pytest-django==4.7.0
pytest-asyncio==0.23.5
freezegun==1.4.0